Rearrange processors into different files
