+++ mailanalysis/__main__.py
@@ -3,12 +3,11 @@ import webbrowser
 from pathlib import Path
 
 import mailanalysis
-from mailanalysis.processors import (
-    ActivityOverTime,
-    DowHourHeatmap,
-    MostFrequentAddresses,
-    ReportHeader,
-)
+from mailanalysis.processors import ReportHeader
+from mailanalysis.processors.activity_over_time import ActivityOverTime
+from mailanalysis.processors.dow_hour_heatmap import DowHourHeatmap
+from mailanalysis.processors.most_frequent_addresses import \
+    MostFrequentAddresses
 
 REPORT_FILE = "report_mail.html"
